Refactor rest_api.js
